% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Breadcrumbs}
\alias{Breadcrumbs}
\title{Breadcrumbs}
\usage{
Breadcrumbs(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/breadcrumbs}
}
\examples{
library(shiny.blueprint)
library(shiny)

items <- list(
  list(href = "/", icon = "folder-close", text = "Users"),
  list(href = "/", icon = "folder-close", text = "Janet"),
  list(icon = "document", text = "image.jpg")
)

ui <- function(id) {
  Breadcrumbs(items = items)
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {})
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
