% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Callout}
\alias{Callout}
\title{Callout}
\usage{
Callout(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/callout}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = Callout(
    title = "Visually important content",
    "The component is a simple wrapper around the CSS API",
    " that provides props for modifiers and optional title element.",
    " Any additional HTML props will be spread to the rendered ", Code("div"), " element."
  ),
  server = function(input, output) {}
)
}
