% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Collapse}
\alias{Collapse}
\title{Collapse}
\usage{
Collapse(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/collapse}
}
\examples{
library(shiny)
library(appsilon.blueprint)

logs <- Pre(
  "[11:53:30] Finished 'typescript-bundle-blueprint' after 769 ms\n",
  "[11:53:30] Starting 'typescript-typings-blueprint'...\n",
  "[11:53:30] Finished 'typescript-typings-blueprint' after 198 ms\n",
  "[11:53:30] write ./blueprint.css\n",
  "[11:53:30] Finished 'sass-compile-blueprint' after 2.84 s\n"
)

if (interactive()) shinyApp(
  ui = tagList(
    Button.shinyInput("toggle", "Toggle logs"),
    reactOutput("ui")
  ),
  server = function(input, output) {
    show <- reactiveVal(FALSE)
    observeEvent(input$toggle, show(!show()))
    output$ui <- renderReact({
      Collapse(isOpen = show(), logs)
    })
  }
)
}
