% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component}
\alias{component}
\title{Documentation template}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{value}{Initial value.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Documentation template
}
\keyword{internal}
