% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ControlGroup}
\alias{ControlGroup}
\title{ControlGroup}
\usage{
ControlGroup(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/control-group}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = ControlGroup(
    fill = FALSE,
    vertical = FALSE,
    HTMLSelect(options = rownames(mtcars)),
    InputGroup(placeholder = "Find car..."),
    Button(icon = "arrow-right"),
  ),
  server = function(input, output) {}
)
}
