% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{EditableText}
\alias{EditableText}
\alias{EditableText.shinyInput}
\title{Editable text}
\usage{
EditableText(...)

EditableText.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/editable-text}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    H2(EditableText(onChange = setInput("header"))),
    EditableText.shinyInput(
      inputId = "body",
      multiline = TRUE,
      minLines = 3, maxLines = 12
    ),
    textOutput("headerValue"),
    textOutput("bodyValue")
  ),
  server = function(input, output) {
    output$headerValue <- renderText(paste("Header:", deparse(input$header)))
    output$bodyValue <- renderText(paste("Body:", deparse(input$body)))
  }
)
}
