% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Icon}
\alias{Icon}
\title{Icon}
\usage{
Icon(...)
}
\description{
Icon
}
\examples{
library(shiny)
library(appsilon.blueprint)

shinyApp(
  ui = tagList(
    Icon(icon = "cross"),
    Icon(icon = "globe", size = 20),
  ),
  server = function(input, output) {}
)
}
