% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Label}
\alias{Label}
\title{Label}
\usage{
Label(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/label}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = Label(
    "Label",
    tags$input(class = "bp4-input")
  ),
  server = function(input, output) {}
)
}
\seealso{
Other HTML elements: 
\code{\link{HTMLTable}()},
\code{\link{html_elements}}
}
\concept{HTML elements}
