% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectComponents.R
\name{MultiSelect2}
\alias{MultiSelect2}
\title{MultiSelect2}
\usage{
MultiSelect2(inputId, items, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{items}{A list of options (containing \code{key}, \code{label} and \code{rightLabel} keys)}

\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#select/multi-select2}
}
\examples{
library(shiny)
library(appsilon.blueprint)

top5Films <- list(
  list(
    label = "The Shawshank Redemption",
    rightLabel = 1994,
    key = 1
  ),
  list(
    label = "The Godfather",
    rightLabel = 1972,
    key = 2
  ),
  list(
    label = "The Godfather: Part II",
    rightLabel = 1974,
    key = 3
  ),
  list(
    label = "The Dark Knight",
    rightLabel = 2008,
    key = 4,
    selected = TRUE
  ),
  list(
    label = "12 Angry Men",
    rightLabel = 1957,
    key = 5,
    selected = TRUE
  )
)

ui <- function(id) {
  ns <- NS(id)
  tagList(
    H3("Multiselect"),
    MultiSelect(
      items = top5Films,
      inputId = ns("values1"),
      tagInputProps = list(
        tagProps = list(
          intent = "danger"
        )
      ),
      noResults = MenuItem(
        disabled = TRUE,
        text = "No results"
      )
    ),
    uiOutput(ns("values1Output")),
    H3("Multiselect2"),
    MultiSelect2(
      items = top5Films,
      inputId = ns("values2"),
      noResults = MenuItem(
        disabled = TRUE,
        text = "No results"
      )
    ),
    uiOutput(ns("values2Output"))
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$values1Output <- renderText({
      purrr::map_chr(input$values1[[1]], ~ .x$label)
    })

    output$values2Output <- renderText({
      purrr::map_chr(input$values2[[1]], ~ .x$label)
    })
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
