% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Navbar}
\alias{Navbar}
\alias{NavbarGroup}
\alias{NavbarHeading}
\alias{NavbarDivider}
\title{Navbar}
\usage{
Navbar(...)

NavbarGroup(...)

NavbarHeading(...)

NavbarDivider(...)
}
\description{
Navbar
}
\examples{
library(shiny)
library(appsilon.blueprint)

shinyApp(
  ui = tagList(
    Navbar(
      NavbarGroup(
        NavbarHeading("Blueprint"),
        NavbarDivider(),
        Button(minimal = "true", icon = "home", text = "Home"),
        Button(minimal = "true", icon = "document", text = "Files")
      ),
      NavbarGroup(
        align = "right",
        Button(minimal = "true", icon = "user"),
        Button(minimal = "true", icon = "refresh")
      )
    )
  ),
  server = function(input, output) {}
)
}
