% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{NonIdealState}
\alias{NonIdealState}
\title{Non-ideal state}
\usage{
NonIdealState(...)
}
\description{
Non-ideal state
}
\examples{
library(shiny)
library(appsilon.blueprint)

shinyApp(
  ui = tagList(
    NonIdealState(
      icon = "search",
      title = "No search results",
      description = Card("Your search didn't match any files.\nTry searching for something else."),
      action = Button(text = "test", icon = "refresh")
    )
  ),
  server = function(input, output) {}
)
}
