% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{NumericInput}
\alias{NumericInput}
\alias{NumericInput.shinyInput}
\title{NumericInput}
\usage{
NumericInput(...)

NumericInput.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/numeric-input}
}
\examples{
library(shiny)
library(appsilon.blueprint)

ui <- function(id) {
  ns <- NS(id)
  tagList(
    H4("Uncontrolled"),
    NumericInput(
      onValueChange = setInput(ns("uncontrolledNumericInput")),
      intent = "primary"
    ),
    textOutput(ns("uncontrolledNumericInputOutput")),
    H4("Controlled"),
    NumericInput.shinyInput(
      inputId = ns("controlledNumericInput"),
      intent = "primary"
    ),
    textOutput(ns("controlledNumericInputOutput"))
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$uncontrolledNumericInputOutput <- renderText(input$uncontrolledNumericInput)
    output$controlledNumericInputOutput <- renderText(input$controlledNumericInput)
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
