% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Popover}
\alias{Popover}
\title{Popover}
\usage{
Popover(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/popover}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = reactOutput("ui"),
  server = function(input, output) {
    isOpen <- reactiveVal(FALSE)
    observeEvent(input$hello, isOpen(TRUE))
    observeEvent(input$dismiss, isOpen(FALSE))

    output$ui <- renderReact({
      Popover(
        isOpen = isOpen(),
        target = Button.shinyInput("hello", "Say Hello", intent = "primary"),
        content = tags$div(
          style = "padding: 1em",
          H5("Hello!"),
          tags$p("Please read this message."),
          Button.shinyInput("dismiss", "Dismiss")
        )
      )
    })
  }
)
}
