% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Radio}
\alias{Radio}
\alias{RadioGroup}
\alias{RadioGroup.shinyInput}
\title{Radio}
\usage{
Radio(...)

RadioGroup(...)

RadioGroup.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/radio}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    H3("Favorite animal"),
    RadioGroup.shinyInput(
      inputId = "animal",
      value = "dog",
      Radio(label = "Cat", value = "cat"),
      Radio(label = "Dog", value = "dog")
    ),
    textOutput("favoriteAnimal"),
    H3("Favorite fruit"),
    reactOutput("fruitRadio"),
    textOutput("favoriteFruit")
  ),
  server = function(input, output) {
    output$favoriteAnimal <- renderText(deparse(input$animal))

    fruit <- reactiveVal()
    observeEvent(input$fruit, fruit(input$fruit))
    output$fruitRadio <- renderReact({
      RadioGroup(
        onChange = JS("event => Shiny.setInputValue('fruit', event.currentTarget.value)"),
        selectedValue = fruit(),
        Radio(label = "Apple", value = "a"),
        Radio(label = "Banana", value = "b"),
        Radio(label = "Cherry", value = "c")
      )
    })
    output$favoriteFruit <- renderText(deparse(fruit()))
  }
)
}
