% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{RangeSlider}
\alias{RangeSlider}
\alias{RangeSlider.shinyInput}
\title{RangeSlider}
\usage{
RangeSlider(...)

RangeSlider.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/sliders.range-slider}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    H4("Uncontrolled"),
    # The uncontrolled slider value change works correctly but the component is not reactive
    Slider(
      min = 0,
      max = 10,
      value = 5,
      stepSize = 0.1,
      labelStepSize = 10,
      onChange = setInput("uncontrolledSlider")
    ),
    textOutput("uncontrolledSliderOutput"),
    H4("Controlled"),
    Slider.shinyInput(
      inputId = "controlledSlider",
      min = 0,
      max = 10,
      stepSize = 0.1,
      labelStepSize = 10
    ),
    textOutput("controlledSliderOutput")
  ),
  server = function(input, output) {
    output$uncontrolledSliderOutput <- renderText(input$uncontrolledSlider)
    output$controlledSliderOutput <- renderText(input$controlledSlider)
  }
)
}
