% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ResizeSensor}
\alias{ResizeSensor}
\title{Resize sensor}
\usage{
ResizeSensor(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/resize-sensor}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    tags$style("
      .resizable {
        overflow: auto;
        resize: both;
        width: 100px;
        height: 100px;
        background: silver;
      }
    "),
    ResizeSensor(
      onResize = JS("entries => Shiny.setInputValue('resize', entries[0].contentRect)"),
      div(
        class = "resizable",
        textOutput("size")
      )
    )
  ),
  server = function(input, output) {
    output$size <- renderText({
      content <- req(input$resize)
      paste0(content$width, "x", content$height)
    })
  }
)
}
