% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run example}
\usage{
runExample(example = NULL, ...)
}
\arguments{
\item{example}{The name of the example to run, or \code{NULL} to retrieve the list of examples.}

\item{...}{Additional arguments to pass to \code{shiny::runApp()}.}
}
\description{
Launch a Shiny example app or list the available examples.
Use \code{appsilon.blueprint::runExample("showcase")} to run a showcase app with all the components.
}
