% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectComponents.R
\name{Select}
\alias{Select}
\title{Select}
\usage{
Select(inputId, items, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{items}{A list of options (containing \code{key}, \code{label} and \code{rightLabel} keys)}

\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#select/select}
}
\examples{
library(shiny)
library(appsilon.blueprint)

top5Films <- list(
  list(
    label = "The Shawshank Redemption",
    rightLabel = 1994,
    key = 1
  ),
  list(
    label = "The Godfather",
    rightLabel = 1972,
    key = 2
  ),
  list(
    label = "The Godfather: Part II",
    rightLabel = 1974,
    key = 3
  ),
  list(
    label = "The Dark Knight",
    rightLabel = 2008,
    key = 4
  ),
  list(
    label = "12 Angry Men",
    rightLabel = 1957,
    key = 5
  )
)

ui <- function(id) {
  ns <- NS(id)
  tagList(
    H3("Select"),
    Select(
      items = top5Films,
      inputId = ns("value1"),
      noResults = MenuItem(
        disabled = TRUE,
        text = "No results"
      )
    ),
    uiOutput(ns("value1Output")),
    H3("Select2"),
    Select2(
      items = top5Films,
      inputId = ns("value2"),
      noResults = MenuItem(
        disabled = TRUE,
        text = "No results"
      )
    ),
    uiOutput(ns("value2Output"))
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$value1Output <- renderText(input$value1$label)
    output$value2Output <- renderText(input$value2$label)
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
