% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setObservable.R
\name{setObservable}
\alias{setObservable}
\title{Changes the state of the app (client-side)}
\usage{
setObservable(..., eventIdentifier = "\\\\(event\\\\)")
}
\description{
Changes the state of the app (client-side)
}
