% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Spinner}
\alias{Spinner}
\title{Spinner}
\usage{
Spinner(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/spinner}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = Spinner(intent = "primary", size = 100),
  server = function(input, output) {}
)
}
