% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Suggest}
\alias{Suggest}
\alias{Suggest.shinyInput}
\title{Suggest}
\usage{
Suggest(...)

Suggest.shinyInput(
  inputId,
  items,
  selected = NULL,
  noResults = "No results.",
  ...
)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{items}{A list of options (character vector or list containing \code{text} and \code{label} entries)}

\item{selected}{Initialy selected item}

\item{noResults}{Message when no results were found}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#select/suggest2}
}
\examples{
library(shiny)
library(appsilon.blueprint)

top5Films <- list(
  list(
    text = "The Shawshank Redemption",
    label = 1994
  ),
  list(
    text = "The Godfather",
    label = 1972
  ),
  list(
    text = "The Godfather: Part II",
    label = 1974
  ),
  list(
    text = "The Dark Knight",
    label = 2008
  ),
  list(
    text = "12 Angry Men",
    label = 1957
  )
)


ui <- function(id) {
  ns <- NS(id)
  tagList(
    H3("Suggest"),
    Suggest.shinyInput(
      inputId = ns("suggest"),
      items = paste("Option", LETTERS),
      inputProps = list(
        placeholder = "Search with Suggest..."
      )
    ),
    uiOutput(ns("suggest_output")),
    H3("Suggest with labels"),
    Suggest.shinyInput(
      inputId = ns("suggest_lab"),
      items = top5Films,
      noResults = "No suggestions."
    ),
    uiOutput(ns("suggest_lab_output"))
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$suggest_output <- renderText(input$suggest$text)
    output$suggest_lab_output <- renderText(input$suggest_lab$text)
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
