% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Switch}
\alias{Switch}
\alias{Switch.shinyInput}
\title{Switch}
\usage{
Switch(...)

Switch.shinyInput(inputId, ..., value = defaultValue)
}
\description{
Switch
}
\examples{
library(shiny)
library(appsilon.blueprint)

withDefault <- function(x, default) {
  if (is.null(x)) default
  else x
}

shinyApp(
  ui = tagList(
    Switch(
      onChange = JS("(event) => Shiny.setInputValue('animate', event.target.checked)"),
      defaultChecked = TRUE,
      label = "Animate"
    ),
    reactOutput("progress")
  ),
  server = function(input, output) {
    output$progress <- renderReact({
      ProgressBar(animate = withDefault(input$animate, TRUE))
    })
  }
)
library(shiny)
library(appsilon.blueprint)

shinyApp(
  ui = tagList(
    Switch.shinyInput(
      inputId = "animate",
      value = TRUE,
      label = "Animate"
    ),
    reactOutput("progress")
  ),
  server = function(input, output) {
    output$progress <- renderReact({
      ProgressBar(animate = input$animate)
    })
  }
)
}
