% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerBrowserEvent.R
\name{triggerBrowserEvent}
\alias{triggerBrowserEvent}
\title{Trigger browser event}
\usage{
triggerBrowserEvent(inputId, props, eventIdentifier = "\\\\(event\\\\)")
}
\description{
Trigger browser event
}
