% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ButtonGroup}
\alias{ButtonGroup}
\title{Button group}
\usage{
ButtonGroup(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/button-group}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = ButtonGroup(
    Button(icon = "database", "Queries"),
    Button(icon = "function", "Functions"),
    AnchorButton(rightIcon = "caret-down", "Options")
  ),
  server = function(input, output) {}
)
}
