% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Checkbox}
\alias{Checkbox}
\alias{Checkbox.shinyInput}
\title{Checkbox}
\usage{
Checkbox(...)

Checkbox.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/checkbox}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    Checkbox(
      onChange = JS("(event) => Shiny.setInputValue('apples', event.target.checked)"),
      defaultChecked = TRUE,
      label = "Apples"
    ),
    Checkbox.shinyInput(
      inputId = "bananas",
      value = TRUE,
      label = "Bananas"
    ),
    textOutput("applesEnabled"),
    textOutput("bananasEnabled")
  ),
  server = function(input, output) {
    output$applesEnabled <- renderText(paste("Apples:", deparse(input$apples)))
    output$bananasEnabled <- renderText(paste("Bananas:", deparse(input$bananas)))
  }
)
}
