% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ControlGroup}
\alias{ControlGroup}
\title{Control group}
\usage{
ControlGroup(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/control-group}
}
\examples{
library(react.blueprint)
library(shiny)

ui <- function(id) {
  ControlGroup(
    HTMLSelect(options = rownames(mtcars)),
    InputGroup(placeholder = "Find car..."),
    Button(icon = "arrow-right"),
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {})
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
