% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{HTMLSelect}
\alias{HTMLSelect}
\alias{HTMLSelect.shinyInput}
\title{HTMLSelect}
\usage{
HTMLSelect(...)

HTMLSelect.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/html-select}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    H4("Uncontrolled"),
    HTMLSelect(
      options = rownames(mtcars),
       onChange = JS("(event) => Shiny.setInputValue('uncontrolledHtmlSelect', event.target.value)")
    ),
    textOutput("uncontrolledHtmlSelectOutput"),
    H4("Controlled"),
    HTMLSelect.shinyInput(
      inputId = "controlledHtmlSelect",
      options = rownames(mtcars)
    ),
    textOutput("controlledHtmlSelectOutput")
  ),
  server = function(input, output) {
    output$uncontrolledHtmlSelectOutput <- renderText(input$uncontrolledHtmlSelect)
    output$controlledHtmlSelectOutput <- renderText(input$controlledHtmlSelect)
  }
)
}
