% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Icon}
\alias{Icon}
\title{Icon}
\usage{
Icon(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/icon}
}
\details{
A list of available icons: \url{https://blueprintjs.com/docs/#icons}
}
\examples{
library(react.blueprint)
library(shiny)

ui <- function(id) {
  tagList(
    Icon(icon = "cross"),
    Icon(icon = "globe", size = 20),
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {})
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
