% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Navbar}
\alias{Navbar}
\alias{NavbarGroup}
\alias{NavbarHeading}
\alias{NavbarDivider}
\title{Navbar}
\usage{
Navbar(...)

NavbarGroup(...)

NavbarHeading(...)

NavbarDivider(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/navbar}
}
\examples{
library(shiny.blueprint)
library(shiny)

ui <- function(id) {
  Navbar(
    NavbarGroup(
      NavbarHeading("Blueprint"),
      NavbarDivider(),
      Button(minimal = TRUE, icon = "home", text = "Home"),
      Button(minimal = TRUE, icon = "document", text = "Files")
    ),
    NavbarGroup(
      align = "right",
      Button(minimal = TRUE, icon = "user"),
      Button(minimal = TRUE, icon = "refresh")
    )
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {})
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
