% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{NonIdealState}
\alias{NonIdealState}
\title{Non-ideal state}
\usage{
NonIdealState(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/non-ideal-state}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = NonIdealState(
    icon = "search",
    title = "No search results",
    description = Card(
      "Your search didn't match any files.",
      tags$br(),
      "Try searching for something else, or create a new file."
    ),
    action = Button(icon = "plus", text = "New file", intent = "primary", outlined = TRUE)
  ),
  server = function(input, output) {}
)
}
