% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/handlers.R
\name{PanelStack}
\alias{PanelStack}
\alias{PanelStack.shinyWrapper}
\alias{openPanel}
\alias{closePanel}
\title{Panel stack (v2)}
\usage{
PanelStack(...)

PanelStack.shinyWrapper(panels, ns = "ps", size = c(300, 250), ...)

openPanel(panelId, ns = "ps")

closePanel(ns = "ps")
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{panels}{List of lists - each list contains \code{title} (string) and \code{content} (HTML)}

\item{ns}{Namespace of given panel stack (required if there's more than 1 panel stack)}

\item{size}{Numeric vector of length 2 - \code{c(width, height)}}

\item{panelId}{Id of the panel to be closed}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/panel-stack2}
}
