% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ProgressBar}
\alias{ProgressBar}
\title{Progress bar}
\usage{
ProgressBar(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/progress-bar}
}
\examples{
library(shiny.blueprint)
library(shiny)

ui <- function(id) {
  ProgressBar(animate = TRUE)
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {})
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
