% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{RangeSlider}
\alias{RangeSlider}
\alias{RangeSlider.shinyInput}
\alias{updateRangeSlider.shinyInput}
\title{Range slider}
\usage{
RangeSlider(...)

RangeSlider.shinyInput(inputId, ..., value = defaultValue)

updateRangeSlider.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{Initial value.}

\item{session}{The Shiny session object.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/sliders.range-slider}
}
\examples{
library(shiny)
library(shiny.blueprint)

ui <- function(id) {
  ns <- NS(id)
  tagList(
    Slider.shinyInput(
      inputId = ns("value"),
      min = 0,
      max = 10,
      stepSize = 0.1,
      labelStepSize = 10
    ),
    textOutput(ns("valueOutput"))
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$valueOutput <- renderText(input$value)
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
