% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ResizeSensor}
\alias{ResizeSensor}
\title{Resize sensor}
\usage{
ResizeSensor(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/resize-sensor}
}
\examples{
library(appsilon.blueprint)
library(shiny)

setInput <- function(inputId, accessor = NULL) {
  # Use ShinyProxy to avoid an exception on app startup:
  # https://github.com/Appsilon/shiny.react/blob/e86d7372e1946dd86dc0ee69e3b4e9ef7f20d7db/js/src/react/ShinyProxy.js#L3 # nolint
  JS(paste0(
    "x => jsmodule['@/shiny.react'].ShinyProxy.setInputValue",
    "('", inputId, "', x", accessor, ")"
  ))
}

ui <- function(id) {
  ns <- NS(id)
  tagList(
    tags$style("
      .resizable {
        overflow: auto;
        resize: both;
        width: 100px;
        height: 100px;
        background: silver;
      }
    "),
    ResizeSensor(
      onResize = setInput(ns("resize"), "[0].contentRect"),
      div(
        class = "resizable",
        textOutput(ns("size"))
      )
    )
  )
}

server <- function(id) {
  moduleServer(id, function(input, output, session) {
    output$size <- renderText({
      content <- req(input$resize)
      paste0(content$width, "x", content$height)
    })
  })
}

if (interactive()) shinyApp(ui("app"), function(input, output) server("app"))
}
