% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TextArea}
\alias{TextArea}
\title{TextArea}
\usage{
TextArea(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for details.}
}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/text-inputs.text-area}
}
\examples{
library(shiny)
library(appsilon.blueprint)

if (interactive()) shinyApp(
  ui = tagList(
    H4("Uncontrolled"),
    TextArea(
      growVertically = TRUE,
      onChange = JS("event => Shiny.setInputValue('uncontrolledTextarea', event.target.value)"),
      large = TRUE,
      intent = "primary"
    ),
    textOutput("uncontrolledTextareaOutput"),
    H4("Controlled"),
    TextArea.shinyInput(
      inputId = "controlledTextarea",
      growVertically = TRUE,
      large = TRUE,
      intent = "primary"
    ),
    textOutput("controlledTextareaOutput")
  ),
  server = function(input, output) {
    output$uncontrolledTextareaOutput <- renderText(input$uncontrolledTextarea)
    output$controlledTextareaOutput <- renderText(input$controlledTextarea)
  }
)
}
