% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toaster.R
\name{Toaster}
\alias{Toaster}
\title{Toaster}
\description{
Documentation: \url{https://blueprintjs.com/docs/#core/components/toast}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Toaster-new}{\code{Toaster$new()}}
\item \href{#method-Toaster-show}{\code{Toaster$show()}}
\item \href{#method-Toaster-clear}{\code{Toaster$clear()}}
\item \href{#method-Toaster-dismiss}{\code{Toaster$dismiss()}}
\item \href{#method-Toaster-clone}{\code{Toaster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toaster-new"></a>}}
\if{latex}{\out{\hypertarget{method-Toaster-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toaster$new(
  toasterId = incrementToasterId(),
  session = shiny::getDefaultReactiveDomain(),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toaster-show"></a>}}
\if{latex}{\out{\hypertarget{method-Toaster-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toaster$show(..., key = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toaster-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Toaster-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toaster$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toaster-dismiss"></a>}}
\if{latex}{\out{\hypertarget{method-Toaster-dismiss}{}}}
\subsection{Method \code{dismiss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toaster$dismiss(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Toaster-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Toaster-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toaster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
